Move integrated timer queue into time-queue-driver
