+++ embassy-executor/src/raw/util.rs
@@ -54,9 +54,4 @@ impl<T> SyncUnsafeCell<T> {
     {
         *self.value.get()
     }
-
-    #[cfg(feature = "integrated-timers")]
-    pub unsafe fn replace(&self, value: T) -> T {
-        core::mem::replace(&mut *self.value.get(), value)
-    }
 }
